+++ app/assets/svg/errorInfoIcon/ErrorInfoIcon.styles.ts
@@ -1,0 +1,12 @@
+import { StyleSheet, ViewStyle } from 'react-native';
+
+interface IErrorInfoIconStyles {
+  icon: ViewStyle;
+}
+
+export const styles = StyleSheet.create<IErrorInfoIconStyles>({
+  icon: {
+    paddingRight: 16,
+    justifyContent: 'center',
+  },
+});
